Add @types/listr to apollo package
I ran into issues with referencing the apollo package from a typescript
file, since the listr types are not present in the package.json on
publish. Unfortunately lerna doesn't ensure that all used dependencies
show up in the package.json on publish. Maybe there is a flag we can use
to ensure that we don't run into these sort of annoying issues in the
future.